requirements.txt
transformers==4.38.2
datasets==2.16.1
torch==2.1.0 
accelerate==0.27.2
tokenizers==0.15.2
nltk==3.8.1
spacy==3.7.2
vaderSentiment==3.3.2
pandas==2.2.1
scikit-learn==1.3.2
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.20.0
streamlit==1.32.2
python-dotenv==1.0.1
pdfkit==1.0.0
jinja2==3.1.3